requirements.txt
discord.py
flask
google-auth-oauthlib
requests
python-dotenv
gunicorn